web_app/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
openai==2.5.0
python-dotenv==1.1.1
fastmcp==2.12.5
httpx==0.28.1
python-multipart==0.0.9
aiohttp==3.9.1